Add collection of longest strings
Instead of using arrays, I find it practical to have small collection classes,
as the interpreter/an IDE does not have to guess what an array contains.
Also, this collection is able to instantiate an amount of LongestString,
which spares you some typework if you need more of them (which is likely for
it's intended purpose).